frontend/requirements.txt
Django>=1.7
psycopg2
django-ratelimit
